// amplify/backend/function/scriptGraphQL/src/API.ts
export type ModelStringInput = {
  ne?: string | null,
  eq?: string | null,
  le?: string | null,
  lt?: string | null,
  ge?: string | null,
  gt?: string | null,
  contains?: string | null,
  notContains?: string | null,
  between?: Array< string | null > | null,
  beginsWith?: string | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
  size?: ModelSizeInput | null,
};
// ...
export enum EmployeeType {
  OTHER = "OTHER",
  NORMAL = "NORMAL",
  MANAGER = "MANAGER",
  SUPER_MANAGER = "SUPER_MANAGER",
}


export enum BooleanType {
  TRUE = "TRUE",
  FALSE = "FALSE",
}


export type ModelEmployeeConditionInput = {
  localID?: ModelIDInput | null,
  employeeGroupLocalId?: ModelIDInput | null,
  superiorUsername?: ModelIDInput | null,
  firstName?: ModelStringInput | null,
  lastName?: ModelStringInput | null,
  grade?: ModelStringInput | null,
  email?: ModelStringInput | null,
  sub?: ModelStringInput | null,
  manager?: ModelEmployeeTypeInput | null,
  isCompanyAdmin?: ModelBooleanInput | null,
  isDeleted?: ModelBooleanTypeInput | null,
  and?: Array< ModelEmployeeConditionInput | null > | null,
  or?: Array< ModelEmployeeConditionInput | null > | null,
  not?: ModelEmployeeConditionInput | null,
};

export type ModelIDInput = {
  ne?: string | null,
  eq?: string | null,
  le?: string | null,
  lt?: string | null,
  ge?: string | null,
  gt?: string | null,
  contains?: string | null,
  notContains?: string | null,
  between?: Array< string | null > | null,
  beginsWith?: string | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
  size?: ModelSizeInput | null,
};

export type ModelEmployeeTypeInput = {
  eq?: EmployeeType | null,
  ne?: EmployeeType | null,
};

export type ModelBooleanInput = {
  ne?: boolean | null,
  eq?: boolean | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
};

export type ModelBooleanTypeInput = {
  eq?: BooleanType | null,
  ne?: BooleanType | null,
};
// ...
export type UpdateEmployeeInput = {
  companyID: string,
  username: string,
  localID?: string | null,
  employeeGroupLocalId?: string | null,
  superiorUsername?: string | null,
  firstName?: string | null,
  lastName?: string | null,
  grade?: string | null,
  email?: string | null,
  sub?: string | null,
  manager?: EmployeeType | null,
  isCompanyAdmin?: boolean | null,
  isDeleted?: BooleanType | null,
};
// ...
export type UpdateEmployeeMutationVariables = {
  input?: UpdateEmployeeInput,
  condition?: ModelEmployeeConditionInput | null,
};